+++ core/main.ts
@@ -95,6 +95,7 @@ export * from "./src/ecs/actions/ToggleVisible"
 export * from "./src/ecs/actions/abilities/IceWall"
 export * from "./src/ecs/actions/attacks/Reload"
 export * from "./src/ecs/actions/attacks/Shoot"
+export * from "./src/ecs/actions/attacks/Whack"
 export * from "./src/ecs/actions/attacks/ZomiAttack"
 export * from "./src/ecs/actions/movement/Point"
 export * from "./src/ecs/actions/movement/Chase"
