+++ core/src/ecs/entities/items/Tools.ts
@@ -1,21 +1,21 @@
 import {
-  Action, Actions, Character, Effects, Entity, Item, KeyMouse,
-  Name, Renderable, SpawnHitbox, SpawnHitboxProps, ValidSounds,
-  XYdifferent, abs, hypot, loadTexture, min, mouseScreen, onHitTeam, playSound, randomInt
+  Actions, Character, Effects, Entity, Item, Name,
+  Renderable, SpawnHitbox, ValidSounds, Whack, XYdifferent,
+  abs, hypot, loadTexture, min, mouseScreen, randomInt
 } from "@piggo-gg/core"
 import { Sprite } from "pixi.js"
 
-export const Axe = (character: Character): Item => {
+export const Tool = (name: string, sound: ValidSounds) => (character: Character): Item => {
 
   let mouseLast = { x: 0, y: 0 }
 
-  const axe: Item = Entity({
-    id: `${character.id}-axe-${randomInt(1000)}`,
+  const tool: Item = Entity({
+    id: `${character.id}-${name}-${randomInt(1000)}`,
     components: {
-      name: Name("axe"),
+      name: Name(name),
       position: character.components.position,
       actions: Actions<any>({
-        "mb1": Whack("thud", 25),
+        "mb1": Whack(sound, 25),
         "spawnHitbox": SpawnHitbox
       }),
       effects: Effects(),
@@ -27,7 +27,7 @@ export const Axe = (character: Character): Item => {
         interpolate: true,
         visible: false,
         rotates: true,
-        dynamic: (_, r, e, w) => {
+        dynamic: (_, r) => {
           const { pointingDelta, rotation } = character.components.position.data
 
           if (rotation) {
@@ -53,7 +53,7 @@ export const Axe = (character: Character): Item => {
           mouseLast = mouseScreen
         },
         setup: async (r: Renderable) => {
-          const textures = await loadTexture("axe.json")
+          const textures = await loadTexture(`${name}.json`)
 
           r.c = new Sprite(textures["0"])
 
@@ -62,105 +62,9 @@ export const Axe = (character: Character): Item => {
       })
     }
   })
-  return axe
+  return tool
 }
 
-export const Pickaxe = (character: Character): Item => {
-
-  let mouseLast = { x: 0, y: 0 }
-
-  const pickaxe: Item = Entity({
-    id: `${character.id}-pickaxe-${randomInt(1000)}`,
-    components: {
-      name: Name("pickaxe"),
-      position: character.components.position,
-      actions: Actions<any>({
-        "mb1": Whack("clink", 10),
-        "spawnHitbox": SpawnHitbox
-      }),
-      effects: Effects(),
-      renderable: Renderable({
-        scaleMode: "nearest",
-        zIndex: character.components.renderable.zIndex,
-        scale: 2.5,
-        anchor: { x: 0.5, y: 0.5 },
-        interpolate: true,
-        visible: false,
-        rotates: true,
-        dynamic: (_, r, e, w) => {
-          const { pointingDelta, rotation } = character.components.position.data
-
-          if (rotation) {
-            character.components.position.rotateDown(rotation > 0 ? 0.1 : -0.1, true)
-          }
-
-          if (XYdifferent(mouseScreen, mouseLast)) {
-
-            const hypotenuse = hypot(pointingDelta.x, pointingDelta.y)
-
-            const hyp_x = pointingDelta.x / hypotenuse
-            const hyp_y = pointingDelta.y / hypotenuse
-
-            r.position = {
-              x: hyp_x * min(20, abs(pointingDelta.x)),
-              y: hyp_y * min(20, abs(pointingDelta.y)) - 5
-            }
-
-            const flip = pointingDelta.x > 0 ? 1 : -1
-            r.setScale({ x: flip, y: 1 })
-          }
-
-          mouseLast = mouseScreen
-        },
-        setup: async (r: Renderable) => {
-          const textures = await loadTexture("pickaxe.json")
-
-          r.c = new Sprite(textures["0"])
-
-          r.setOutline(0x000000)
-        }
-      })
-    }
-  })
-  return pickaxe
-}
-
-const Whack = (sound: ValidSounds, damage: number) => Action<KeyMouse & { character: Character }>(({ world, params, entity }) => {
-  if (!entity) return
-
-  const { mouse, character } = params
-
-  if (!mouse || !character) return
-
-  const { position } = entity.components
-  if (!position) return
-
-  if (position.data.pointingDelta.x > 0) {
-    position.rotateUp(1)
-  } else {
-    position.rotateDown(1)
-  }
-
-  const angle = Math.atan2(position.data.pointingDelta.y, position.data.pointingDelta.x)
-
-  const hurtboxParams: SpawnHitboxProps = {
-    pos: {
-      x: position.data.x + Math.cos(angle) * 30,
-      y: position.data.y + Math.sin(angle) * 30,
-    },
-    team: character.components.team,
-    radius: 20,
-    damage,
-    id: randomInt(1000),
-    visible: false,
-    expireTicks: 2,
-    onHit: () => {
-      playSound(world.client?.sounds[sound])
-    },
-    onExpire: () => {
-      playSound(world.client?.sounds["whiff"])
-    }
-  }
-
-  world.actionBuffer.push(world.tick + 1, entity.id, { action: "spawnHitbox", params: hurtboxParams })
-}, 15)
+export const Axe = Tool("axe", "thud")
+export const Sword = Tool("sword", "slash")
+export const Pickaxe = Tool("pickaxe", "clink")
