+++ web/res/sword.json
@@ -1,0 +1,10 @@
+{
+  "frames": {
+    "0": { "frame": { "x": 1, "y": 1, "w": 18, "h": 18 } }
+  },
+  "meta": {
+    "image": "sword.png",
+    "size": { "w": 20, "h": 20 },
+    "scale": "2.5"
+  }
+}
